navext: common function for generating xml ids.
