+++ vec/vec-v2x/src/main/java/com/foursoft/harness/vec/v2x/visitor/DefaultXmlIdGenerator.java
@@ -23,35 +23,34 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.harness.vec.scripting.utils;
+package com.foursoft.harness.vec.v2x.visitor;
 
 import com.foursoft.harness.navext.runtime.model.ModifiableIdentifiable;
 import com.foursoft.harness.vec.common.util.StringUtils;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
-public class XmlIdGenerator {
-
-    private static final String XML_ID_PREFIX = "id_";
+public class DefaultXmlIdGenerator implements XmlIdGenerator {
 
     private final AtomicInteger counter = new AtomicInteger(0);
 
-    public void createIdForXmlBean(ModifiableIdentifiable aBean) {
+    public void createIdForXmlBean(final ModifiableIdentifiable aBean) {
         createIdForXmlBean(aBean, derivePrefix(aBean));
     }
 
-    private void createIdForXmlBean(ModifiableIdentifiable aBean, String prefix) {
+    protected void createIdForXmlBean(final ModifiableIdentifiable aBean, final String prefix) {
         if (StringUtils.isEmpty(aBean.getXmlId())) {
             aBean.setXmlId(generateNewXmlId(prefix));
         }
     }
 
-    private String generateNewXmlId(String prefix) {
+    private String generateNewXmlId(final String prefix) {
         return String.format("%s%05d", prefix, counter.getAndIncrement());
     }
 
-    private String derivePrefix(ModifiableIdentifiable aBean) {
-        return aBean.getClass().getSimpleName().replace("Vec", "") + "_";
+    private String derivePrefix(final ModifiableIdentifiable aBean) {
+        return aBean.getClass()
+                .getSimpleName()
+                .replace("Vec", "") + "_";
     }
-
 }
